refactor(metrics-bridge): extract peg poll cycle coordinator
